se sigue

+++ Models/LocalesProductosInicial.cs
@@ -4,5 +4,5 @@ public class LocalesProductosInicial
     public int Id;
     public int IdLocal;
     public int IdProducto;
-    public float PrecioInicial;
+    public int PrecioInicial;
 }
